+++ exec.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exec_push.c                                        :+:      :+:    :+:   */
+/*   exec.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lenivorb <lenivorb@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/28 12:54:33 by lenivorb          #+#    #+#             */
-/*   Updated: 2026/06/02 18:17:16 by rcollet          ###   ########.fr       */
+/*   Updated: 2026/06/05 14:49:49 by rcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,14 +21,14 @@
 
 // --- define ---
 
-void	exec(t_interface *stacks, t_op_track *tracker, int idx)
+void	exec(t_interface *stacks, int idx)
 {
-	(tracker -> calls)[idx]++;
+	(stacks -> stats -> calls)[idx]++;
 	((stacks -> func)[idx] -> f)(stacks)
-	ft_printf("%s\n", (stacks -> func)[idx] -> name);
+	ft_putendl((stacks -> func)[idx] -> name, 1);
 }
 
-void	exec_n(t_interface *stacks, t_op_track *tracker, int idx, int n)
+void	exec_n(t_interface *stacks, int idx, int n)
 {
 	while (n--)
 		exec(stacks, tracker, idx);