+++ interface_rotate.c
@@ -6,22 +6,22 @@
 /*   By: rcollet <rcollet@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/27 17:19:39 by rcollet           #+#    #+#             */
-/*   Updated: 2026/05/27 17:21:03 by rcollet          ###   ########.fr       */
+/*   Updated: 2026/06/05 14:38:59 by rcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ra(t_interface *stacks)
+void	ra(t_interface *c)
 {
-	r(stacks -> a);
+	r((c -> stacks)[0]);
 }
 
-void	rb(t_interface *stacks)
+void	rb(t_interface *c)
 {
-	r(stacks -> b);
+	r((c -> stacks)[1]);
 }
 
-void	rr(t_interface *stacks)
+void	rr(t_interface *c)
 {
-	r(stacks -> a);
-	r(stacks -> b);
+	ra(c);
+	rb(c);
 }