+++ interface_rrotate.c
@@ -6,22 +6,22 @@
 /*   By: rcollet <rcollet@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/27 17:21:06 by rcollet           #+#    #+#             */
-/*   Updated: 2026/05/27 17:22:24 by rcollet          ###   ########.fr       */
+/*   Updated: 2026/06/05 14:40:10 by rcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	rra(t_interface	*stacks)
+void	rra(t_interface *c)
 {
-	rr(stacks -> a);
+	rr((c -> stacks)[0]);
 }
 
-void	rrb(t_interface *stacks)
+void	rrb(t_interface	*c)
 {
-	rr(stacks -> b);
+	rr((c -> stacks)[1]);
 }
 
-void	rrr(t_interface *stacks)
+void	rrr(t_interface *c)
 {
-	rr(stacks -> a);
-	rr(stacks -> b);
+	rra(c);
+	rrb(c);
 }